requirements.txt
fastapi==0.115.11
aio-pika==9.5.5
python-dotenv==1.0.1
uvicorn==0.34.0
pydantic==2.3.0
pydantic-settings==2.0.3
pika==1.3.2
minio==7.2.0
PyPDF2==3.0.1
python-multipart==0.0.6
anthropic==0.49.0